test: 将 pull_request 相关测试放入 resolve 文件夹下
