+++ tests/plugins/github/resolve/test_remove_pull_request.py
@@ -5,30 +5,12 @@ from nonebug import App
 from pytest_mock import MockerFixture
 
 from tests.plugins.github.event import get_mock_event
+from tests.plugins.github.resolve.utils import get_pr_labels
 from tests.plugins.github.utils import (
     MockIssue,
     generate_issue_body_remove,
     get_github_bot,
 )
-
-
-def get_pr_labels(labels: list[str]):
-    from githubkit.rest import PullRequestPropLabelsItems as Label
-
-    return [
-        Label.model_construct(
-            **{
-                "color": "2A2219",
-                "default": False,
-                "description": "",
-                "id": 2798075966,
-                "name": label,
-                "node_id": "MDU6TGFiZWwyNzk4MDc1OTY2",
-                "url": "https://api.github.com/repos/he0119/action-test/labels/Remove",
-            }
-        )
-        for label in labels
-    ]
 
 
 async def test_remove_process_pull_request(
